Vulnerability: Upgrade `rack-protection` gem
We upgrade the `rack-protection` gem from 1.5.4 → 2.0.1 to remove a
vulnerability which was pointed out in GitHub:

https://nvd.nist.gov/vuln/detail/CVE-2018-1000119

We're taking the opportunity to upgrade the `rack` gem, too. And
`sinatra` as well, to future-proof our tests. Our new test requires a
minimum of ruby 2.2, but that's okay because our ruby buildpack uses
2.4.

[#156815902]

Signed-off-by: Brian Cunnie <bcunnie@pivotal.io>